requirements.txt
fastapi
uvicorn[standard]
kafka-python
mysql-connector-python
pydantic
requests
python-dotenv
